requirements.txt
django==4.1
djangorestframework==3.13.1
python-jose==3.3.0
python-dotenv==1.0.0
celery==5.3.1
redis==4.3.4
psycopg2-binary==2.9.1
environs==8.0.0